+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* In `Down::ChunkedIO` only create the Tempfile if it's going to be used (@janko-m)
+
 ## 3.1.0 (2017-06-16)
 
 * Split `Down::NotFound` into explanatory exceptions (@janko-m)