+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Fix `Down::ChunkedIO` not returning second part of the last chunk if it was previously partially read (@janko-m)
+
 * Strip internal variables from `Down::ChunkedIO#inspect` and show only the important ones (@janko-m)
 
 * Add `Down::ChunkedIO#closed?` (@janko-m)