+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## HEAD
+## 3.2.0 (2017-06-21)
 
 * Add `Down::ChunkedIO#readpartial` for more memory efficient reading (@janko-m)